% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radjust.R
\docType{data}
\name{crohn}
\alias{crohn}
\title{EDIT}
\format{A data frame with 126 rows and 3 columns:
\tabular{lll}{
\code{index} \tab integer \tab just the row number. \cr
\code{pv1}   \tab numeric \tab p-value from study 1. \cr
\code{pv2}   \tab numeric \tab p-value from study 2.
}}
\source{
Barrett, Jeffrey C., et al. "Genome-wide association defines more than 30 distinct susceptibility loci for
 Crohn's disease." Nature genetics 40.8 (2008): 955.
}
\usage{
crohn
}
\description{
EDIT ...
}
\references{
Bogomolov, M. and Heller, R. (2013). Discovering findings that replicate from a primary study of high dimension to a follow-up study.
Journal of the American Statistical Association, Vol. 108, No. 504, Pp. 1480-1492.
}
\seealso{
the example in \code{\link{radjust_pf}} uses this data.
}
\keyword{datasets}
