% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radjust_sym.R
\name{radjust_sym}
\alias{radjust_sym}
\title{Adjust p-values for Replicability across Two Independent Studies with Multiple Endpoints}
\usage{
radjust_sym(pv1, pv2, w1 = 0.5, input_type = c("selected_features",
  "all_features"), general_dependency = FALSE,
  directional_rep_claim = FALSE, selection_method = c("alpha-adjusted",
  "alpha-adaptive", "none"), alpha = if (selection_method == "none") NULL else
  0.05)
}
\arguments{
\item{pv1, pv2}{numeric vectors of p-values from the two studies, corresponding to the selected features (the default) or all the features; can be either of the same length or named.}

\item{w1}{the relative weight for the p-values from study 1.}

\item{input_type}{whether \code{pv1} and \code{pv2} contain all the p-values from the studies or only the selected (the default).}

\item{general_dependency}{whether the correct for general dependency.}

\item{selection_method}{...}

\item{alpha}{the threshold for the selection method ...}
}
\value{
a list with the following elements:
  \tabular{ll}{
    \code{call}  \tab  the function call \cr
    \code{inputs}  \tab  a list with the function's input parameters (except \code{pv1} and \code{pv2}) \cr
    \code{results_table} \tab  a data frame with the features selected in both studies and their r-values. The data frame includes 6 columns: \cr
    \code{selected1, selected2}  \tab the named or indexed features selected in study 1, when \code{selection_method!="none"} \cr
    \code{n_selected1, n_selected2} \tab  the named or indexed features selected in study 1, when \code{selection_method!="none"} \cr
    \code{pi1, pi2}  \tab  the estimate of the true-nulls fraction in the study1\\study2
  }

  \tabular{lll}{
    \code{name}         \tab char.   \tab the name of the feature as extracted from the named vectors, or the location, if the input vectors are not named. \cr
    \code{p.value.1}    \tab numeric \tab the one-sided p-value from study 1 as inputed. In the case of \code{directional_rep_claim==TRUE} the 'real' one-sided p-values is presented, i.e: \code{pmin(pv1,1-pv1}. \cr
    \code{p.value.2}    \tab numeric \tab same as in \code{p.value.1}. \cr
    \code{r.value}      \tab numeric \tab the replicability adjusted p-value (= r-value). \cr
    \code{Direction}    \tab char.   \tab the direction of the replicability claim, when \code{directional_rep_claim==TRUE}. \cr
    \code{Significant}  \tab char.   \tab a notion whether the replicability claim is significant, when \code{selection_method!="none"}.
  }
}
\description{
The function receives two vectors of p-values from two independent studies and adjust
 them to control the False Discovery Rate of replicability claims.
}
\details{
EDIT extended details about the function.
}
\examples{

data(mice)
## transform the two-sided p-values to one-sided in the same direction (left);
## we use the direction of the test statistic to do so and assume that it is continuous
pv1 <- ifelse(mice$dir_is_left1, mice$twosided_pv1/2, 1-mice$twosided_pv1/2)
pv2 <- ifelse(mice$dir_is_left2, mice$twosided_pv2/2, 1-mice$twosided_pv2/2)
## run the examples as in the article

mice_rv_adaptive <- radjust_sym(pv1, pv2, input_type = "all", directional_rep_claim = T, selection_method = "alpha-adaptive")
mice_rv_adjusted <- radjust_sym(pv1, pv2, input_type = "all", directional_rep_claim = T, selection_method = "alpha-adjusted")

mice_rv_no_sel <- radjust_sym(pv1, pv2, input_type = "selected", directional_rep_claim = T, selection_method = "none")
mice_rv_no_sel

print(mice_rv_adaptive)
}
\references{
Add here the reference to the article.
}
\seealso{
\code{\link{radjust_pf}} for replicability analysis in primary and follow-up studies.
}
