% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radjust.R
\docType{data}
\name{crohn}
\alias{crohn}
\title{EDIT}
\format{A data frame with 3 rows and 12X columns:
\tabular{lll}{
\code{index} \tab integer \tab just the row number. \cr    
\code{pv1}   \tab numeric \tab p-value from study 1. \cr    
\code{pv2}   \tab numeric \tab p-value from study 2.
}}
\source{
\url{} <- EDIT optional, Do we have an online source of the data?
}
\usage{
crohn
}
\description{
EDIT ...
}
\references{
Richter, S. Helene, et al. "Effect of population heterogenization on the reproducibility of mouse behavior:
 a multi-laboratory study." PLoS One 6.1 (2011): e16461.
}
\seealso{
the example in \code{\link{radjust_pf}} uses this data.
}
\keyword{datasets}
