% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radjust.R
\docType{data}
\name{mice}
\alias{mice}
\title{p-values of 29 Behavioural Measures in Two Studies}
\format{A data frame with 29 rows and 5 columns:
\tabular{lll}{
\code{feature_name} \tab char.   \tab The name of the measure and the test, concatenated. \cr    
\code{twosided_pv1} \tab numeric \tab the \emph{two-sided} p-value from study 1. \cr    
\code{twosided_pv2} \tab numeric \tab EDIT \cr        
\code{dir_is_left1} \tab logical \tab whether the direction of the test statistic from study 1 is \emph{left}. \cr    
\code{dir_is_left2} \tab logical \tab EDIT
}}
\source{
\url{} <- EDIT optional, Do we have an online source of the data?
}
\usage{
mice
}
\description{
EDIT ... note that the table contains two-sided p-values. To transform all
 the two-sided p-values to one sided in the same direction, see the example in \code{\link{radjust_sym}}.
}
\references{
Barrett, Jeffrey C., et al. "Genome-wide association defines more than 30 distinct susceptibility loci for
 Crohn's disease." Nature genetics 40.8 (2008): 955.
}
\seealso{
\code{\link{radjust_sym}}
}
\keyword{datasets}
