% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radjust_pf.R
\name{radjust_pf}
\alias{radjust_pf}
\title{Adjust p-values for Replicability across (Independent) Primary and Follow-up Studies with Multiple Endpoints}
\usage{
radjust_pf(pv1, pv2, m, c2 = 0.5, l00 = 0, variation = c("none",
  "use_m_star", "use_threshold"), tt = NULL, alpha = 0.05)
}
\arguments{
\item{pv1, pv2}{numeric vectors, of the same length, of the p-values from the from the primary study (\code{pv1}) with
the corresponding p-values from the follow-up study (\code{pv2}).}

\item{m}{the number of features examined in the primary study (can be bigger than \code{length(pv1)}).}

\item{c2}{the relative boost to the p-values from the \strong{primary} study.
\code{c2 = 0.5} (the default) is recommended. It was observed in simulations to yield
similar power to procedure with the optimal value (which is unknown for real data).}

\item{l00}{a lower bound of the fraction of features (out of m) with true null hypotheses in both studies.
For example, for GWAS on the whole genome, the choice of 0.8 is conservative
in typical applications.}

\item{variation}{\describe{
   \item{none}{the default.}
   \item{use_m_star}{use \eqn{m^*=m\sum_{i=1}^{m}\frac{1}{i}}{m*=m*sum(1/i)} modification of \code{m}.}
   \item{use_threshold}{c1 is computed given the threshold \code{tt}.}
}
Both variations guarantee that the procedure that decleares all r-values below \code{alpha} as replicability claims,
controls the FDR at level \code{alpha}, for any type of dependency of the p-values in the primary study.}

\item{tt}{the selection rule threshold for p-values from the primary study; must be supplied when
variation 'use_threshold' is selected, otherwise ignored.}
}
\value{
vector of length of \code{pv2} and \code{pv2}, containing the r-values.
}
\description{
The function computes r-values given two vectors of p-values from primary and
 follow-up studies. The r-values assess the False Discovery Rate (FDR) of repilcability
 claims across the primary and follow-up studies.
}
\details{
EDIT  extended details about the function.
}
\note{
The function is also available as a web applet:  \url{http://www.math.tau.ac.il/~ruheller/App.html}
}
\examples{
 data(crohn)
 rv  <- radjust_pf(pv1 = crohn$pv1, pv2 = crohn$pv1, m = 635547, l00 = 0.8)
 rv2 <- radjust_pf(pv1 = crohn$pv1, pv2 = crohn$pv1, m = 635547, l00 = 0.8, variation="use_threshold",tt = 1e-5)

}
\seealso{
\code{\link{radjust_sym}} for replicability analysis in two symmetric design (EDIT)
}
